requirements.txt
Flask==1.1.2
pandas==1.1.3
Pillow==8.0.0
python-docx==0.8.10
requests==2.24.0
setuptools==50.3.0.post20201006